+++ lib/workingsetcache/cache.go
@@ -62,7 +62,7 @@ type Cache struct {
 	maxBytes int
 
 	// mu serializes access to curr, prev and mode
-	// in expirationWatcher, prevCacheWatcher and cacheSizeWatcher.
+	// in expirationWatcher, prevCacheWatcher,cacheSizeWatcher, UpdateStats and Reset.
 	mu sync.Mutex
 
 	wg     sync.WaitGroup
@@ -118,9 +118,11 @@ func loadWithExpire(filePath string, maxBytes int, expireDuration time.Duration)
 
 	// The cache has been successfully loaded in full.
 	// Set its' mode to `whole`.
-	// There is no need in runWatchers call.
 	prev := fastcache.New(1024)
-	return newCacheInternal(curr, prev, whole, maxBytes)
+	c := newCacheInternal(curr, prev, whole, maxBytes)
+	c.runWatchers(expireDuration)
+
+	return c
 }
 
 // New creates new cache with the given maxBytes capacity and *cacheExpireDuration expiration.
@@ -178,9 +180,9 @@ func (c *Cache) expirationWatcher(expireDuration time.Duration) {
 		}
 		c.mu.Lock()
 		if c.mode.Load() != split {
-			// Stop the expirationWatcher on non-split mode.
+			// Do nothing in non-split mode.
 			c.mu.Unlock()
-			return
+			continue
 		}
 		// Reset prev cache and swap it with the curr cache.
 		prev := c.prev.Load()
@@ -222,7 +224,7 @@ func (c *Cache) prevCacheWatcher() {
 		if c.mode.Load() != split {
 			// Do nothing in non-split mode.
 			c.mu.Unlock()
-			return
+			continue
 		}
 		prev := c.prev.Load()
 		curr := c.curr.Load()
@@ -257,6 +259,7 @@ func (c *Cache) cacheSizeWatcher() {
 	t := time.NewTicker(checkInterval)
 	defer t.Stop()
 
+again:
 	var maxBytesSize uint64
 	for {
 		select {
@@ -265,6 +268,7 @@ func (c *Cache) cacheSizeWatcher() {
 		case <-t.C:
 		}
 		if c.mode.Load() != split {
+			// Do nothing in non-split mode.
 			continue
 		}
 		var cs fastcache.Stats
@@ -309,6 +313,10 @@ func (c *Cache) cacheSizeWatcher() {
 			return
 		case <-t.C:
 		}
+		if c.mode.Load() != switching {
+			// mode was changed by the concurrent goroutine
+			goto again
+		}
 		var cs fastcache.Stats
 		curr := c.curr.Load()
 		curr.UpdateStats(&cs)
@@ -318,6 +326,11 @@ func (c *Cache) cacheSizeWatcher() {
 	}
 
 	c.mu.Lock()
+	if c.mode.Load() != switching {
+		// mode was changed by the concurrent goroutine
+		c.mu.Unlock()
+		goto again
+	}
 	c.mode.Store(whole)
 	prev = c.prev.Load()
 	c.prev.Store(fastcache.New(1024))
@@ -326,6 +339,8 @@ func (c *Cache) cacheSizeWatcher() {
 	updateCacheStatsHistory(&c.csHistory, &cs)
 	prev.Reset()
 	c.mu.Unlock()
+
+	goto again
 }
 
 // Save saves the cache to filePath.
@@ -347,6 +362,9 @@ func (c *Cache) Stop() {
 
 // Reset resets the cache.
 func (c *Cache) Reset() {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
 	var cs fastcache.Stats
 	prev := c.prev.Load()
 	prev.UpdateStats(&cs)
@@ -355,13 +373,23 @@ func (c *Cache) Reset() {
 	curr.UpdateStats(&cs)
 	updateCacheStatsHistory(&c.csHistory, &cs)
 	curr.Reset()
-	// Reset the mode to `split` in the hope the working set size becomes smaller after the reset.
+	// Reset the mode to `split` in the order to properly reset background workers.
+	mode := c.mode.Load()
 	c.mode.Store(split)
 	c.copiedFromPrev.Store(0)
+	if mode != split {
+		// non-split mode changes size of the caches
+		// so we have to restore it into original size of split mode
+		c.prev.Store(fastcache.New(c.maxBytes / 2))
+		c.curr.Store(fastcache.New(c.maxBytes / 2))
+	}
 }
 
 // UpdateStats updates fcs with cache stats.
 func (c *Cache) UpdateStats(fcs *fastcache.Stats) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
 	updateCacheStatsHistory(fcs, &c.csHistory)
 
 	var cs fastcache.Stats